Conflicts:
	web/.idea/workspace.xml
	web/application/views/compras_view.php
	web/application/views/proveedores_view.php
	web/application/views/tarjetas_view.php

+++ web/application/views/maquinas_view.php
@@ -2,11 +2,13 @@
 if (isset($actualizarMaquina)) {
     $ID = '<p><input type="hidden" name="ID" value="' . $this->uri->segment(3) . '"></p>';
     $Descripcion = $actualizarMaquina->Descripcion;
+    $numeroEconomico = $actualizarMaquina->numeroEconomico;
     $action = 'actualizar';
     $button = 'Actualizar';
 } else {
     $ID = '';
     $Descripcion = '';
+    $numeroEconomico = '';
     $action = 'insertar';
     $button = 'Guardar';
 }
@@ -19,13 +21,19 @@ if (isset($actualizarMaquina)) {
                 <h1 class="page-header">Maquinas</h1>
 
                 <div class="divider"></div>
-                <?php if ($this->session->userdata('perfil') == 'Administrador') { ?>
+                <?php if ($this->session->userdata('perfil') == 'Administrador' || $this->session->userdata('perfil') == 'Compras') { ?>
                     <form action="<?php echo base_url(); ?>maquinas/<?php echo $action; ?>" method="post" class="margin-bottom">
                         <?php echo $ID; ?>
-                        <div class="form-group">
+                        <div class="col-sm-6">
                             <label for="Descripcion">Descripcion:</label>
                             <input type="text" class="form-control" id="Descripcion" name="Descripcion"
                                    value="<?php echo $Descripcion ?>"/>
+                            <div class="clearfix"></div>
+                        </div>
+                        <div class="col-sm-6">
+                            <label for="numeroEconomico">Número Económico:</label>
+                            <input type="text" class="form-control" id="numeroEconomico" name="numeroEconomico"
+                                   value="<?php echo $numeroEconomico ?>"/>
                             <div class="clearfix"></div>
                         </div>
                         <div class="form-group">
@@ -46,24 +54,28 @@ if (isset($actualizarMaquina)) {
                             <thead>
                             <th>Maquina</th>
                             <th>Descripción</th>
-                            <?php if ($this->session->userdata('perfil') == 'Administrador') { ?>
+                            <th>Número Económico</th>
+                            <?php if ($this->session->userdata('perfil') == 'Administrador' || $this->session->userdata('perfil') == 'Compras') { ?>
                             <th class="text-center">Editar</th>
+                            <?php if ($this->session->userdata('perfil') == 'Administrador'){?>
                             <th class="text-center">Eliminar</th>
-                            <?php } ?>
+                            <?php } } ?>
                             </thead>
                             <tbody>
                             <?php foreach ($maquinasGuardadas as $maquina) : ?>
                                 <tr>
                                     <td><?php echo $maquina->ID; ?></td>
                                     <td><?php echo $maquina->Descripcion; ?></td>
-                                    <?php if ($this->session->userdata('perfil') == 'Administrador') { ?>
+                                    <td><?php echo $maquina->numeroEconomico; ?></td>
+                                    <?php if ($this->session->userdata('perfil') == 'Administrador' || $this->session->userdata('perfil') == 'Compras') { ?>
                                         <td class="text-center">
                                             <a href="<?php echo base_url(); ?>maquinas/index/<?php echo $maquina->ID; ?>"><i
                                                     class="fa fa-pencil-square-o"></i></a></td>
-                                        <td class="text-center">
+                                        <?php if ($this->session->userdata('perfil') == 'Administrador'){?>
+                                            <td class="text-center">
                                             <a href="<?php echo base_url(); ?>maquinas/eliminar/<?php echo $maquina->ID; ?>"><i
                                                     class="fa fa-times"></i></a></td>
-                                    <?php } ?>
+                                    <?php }} ?>
                                 </tr>
                             <?php endforeach; ?>
                             </tbody>
